Move log tools to dedicated file.
