+++ octoprint_mrbeam/camera/__init__.py
@@ -13,7 +13,7 @@ from octoprint_mrbeam.mrb_logger import mrb_logger
 
 SUCCESS_WRITE_RETVAL = 1
 
-from octoprint_mrbeam.util import logtime, logme
+from octoprint_mrbeam.util.log import logtime, logme
 
 try:
 	import picamera
