+++ octoprint_mrbeam/camera/calibration.py
@@ -13,7 +13,8 @@ import numpy as np
 from copy import copy
 
 from octoprint_mrbeam.mrbeam_events import MrBeamEvents
-from octoprint_mrbeam.util import logme, logtime, logExceptions, makedirs
+from octoprint_mrbeam.util import makedirs
+from octoprint_mrbeam.util.log import logme, logtime, logExceptions, Logger
 from octoprint_mrbeam.mrb_logger import mrb_logger
 
 CB_ROWS = 5
@@ -46,7 +47,7 @@ TMP_RAW_FNAME_RE_NPZ =  'tmp_raw_img_[0-9]+.jpg.npz$'
 # REMOTE_CALIBRATE_EXEC = path.join(REMOTE_CALIBRATION_FOLDER, "calibrate2.py")
 # MY_HOSTNAME = "MrBeam-8ae9"
 
-class BoardDetectorDaemon(Thread):
+class BoardDetectorDaemon(Thread, Logger):
 	"""Processes images of chessboards to calibrate the lens used to take the pictures."""
 
 	def __init__(self,
@@ -58,8 +59,6 @@ class BoardDetectorDaemon(Thread):
 	             event_bus=None,
 		     rawImgLock=None):
 		# runCalibrationAsap : run the lens calibration when we have enough pictures ready
-		self._logger = mrb_logger(__name__ + '.' + self.__class__.__name__, lvl=logging.INFO)
-		self._logger.debug("Initiating the Board Detector Daemon")
 		self.event_bus = event_bus
 		self.rawImgLock = rawImgLock
 
@@ -97,7 +96,8 @@ class BoardDetectorDaemon(Thread):
 		self._startWhenIdle.clear()
 		self.path_inc = 0
 
-		super(self.__class__, self).__init__(target=self.processInputImages, name=self.__class__.__name__)
+		Thread.__init__(self, target=self.processInputImages, name=self.__class__.__name__)
+		Logger.__init__(self, name=__name__ + '.' + self.__class__.__name__)
 
 		# self.daemon = False
 		# catch SIGTERM used by Process.terminate()
